+++ front/lib/actions/mcp_internal_actions/servers/data_sources_file_system.ts
@@ -7,6 +7,7 @@ import type { DataSourcesToolConfigurationType } from "@app/lib/actions/mcp_inte
 import { ConfigurableToolInputSchemas } from "@app/lib/actions/mcp_internal_actions/input_schemas";
 import type {
   FilesystemPathType,
+  SearchQueryResourceType,
   SearchResultResourceType,
 } from "@app/lib/actions/mcp_internal_actions/output_schemas";
 import {
@@ -16,15 +17,16 @@ import {
 } from "@app/lib/actions/mcp_internal_actions/servers/common/find_tags_tool";
 import {
   makeQueryResource,
-  makeQueryResourceForFind,
-  makeQueryResourceForList,
+  renderMimeType,
   renderNode,
   renderSearchResults,
 } from "@app/lib/actions/mcp_internal_actions/servers/search/rendering";
 import {
-  fetchAgentDataSourceConfiguration,
+  getAgentDataSourceConfigurations,
+  makeDataSourceViewFilter,
+} from "@app/lib/actions/mcp_internal_actions/servers/search/utils";
+import {
   getCoreSearchArgs,
-  parseDataSourceConfigurationURI,
   shouldAutoGenerateTags,
 } from "@app/lib/actions/mcp_internal_actions/servers/utils";
 import {
@@ -35,17 +37,14 @@ import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers"
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { actionRefsOffset, getRetrievalTopK } from "@app/lib/actions/utils";
 import { getRefs } from "@app/lib/api/assistant/citations";
-import type { DataSourceConfiguration } from "@app/lib/api/assistant/configuration";
 import config from "@app/lib/api/config";
 import { ROOT_PARENT_ID } from "@app/lib/api/data_source_view";
 import type { InternalMCPServerDefinitionType } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
 import { getDisplayNameForDocument } from "@app/lib/data_sources";
-import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import logger from "@app/logger/logger";
 import type {
-  ConnectorProvider,
   ContentNodeType,
   CoreAPIContentNode,
   CoreAPIError,
@@ -53,12 +52,9 @@ import type {
   Result,
 } from "@app/types";
 import {
-  assertNever,
   CoreAPI,
   DATA_SOURCE_NODE_ID,
   dustManagedCredentials,
-  Err,
-  Ok,
   parseTimeFrame,
   removeNulls,
   stripNullBytes,
@@ -1003,143 +999,6 @@ const createServer = (
   return server;
 };
 
-// Type to represent data source configuration with resolved data source model
-type ResolvedDataSourceConfiguration = DataSourceConfiguration & {
-  dataSource: {
-    dustAPIProjectId: string;
-    dustAPIDataSourceId: string;
-    connectorProvider: ConnectorProvider | null;
-    name: string;
-  };
-};
-
-export async function getAgentDataSourceConfigurations(
-  auth: Authenticator,
-  dataSources: DataSourcesToolConfigurationType
-): Promise<Result<ResolvedDataSourceConfiguration[], Error>> {
-  const configResults = await concurrentExecutor(
-    dataSources,
-    async (dataSourceConfiguration) => {
-      const configInfoRes = parseDataSourceConfigurationURI(
-        dataSourceConfiguration.uri
-      );
-
-      if (configInfoRes.isErr()) {
-        return configInfoRes;
-      }
-
-      const configInfo = configInfoRes.value;
-
-      switch (configInfo.type) {
-        case "database": {
-          // Database configuration
-          const r = await fetchAgentDataSourceConfiguration(configInfo.sId);
-          if (r.isErr()) {
-            return r;
-          }
-          const agentConfig = r.value;
-          const dataSourceViewSId = DataSourceViewResource.modelIdToSId({
-            id: agentConfig.dataSourceView.id,
-            workspaceId: agentConfig.dataSourceView.workspaceId,
-          });
-          const resolved: ResolvedDataSourceConfiguration = {
-            workspaceId: agentConfig.dataSourceView.workspace.sId,
-            dataSourceViewId: dataSourceViewSId,
-            filter: {
-              parents:
-                agentConfig.parentsIn || agentConfig.parentsNotIn
-                  ? {
-                      in: agentConfig.parentsIn || [],
-                      not: agentConfig.parentsNotIn || [],
-                    }
-                  : null,
-              tags:
-                agentConfig.tagsIn || agentConfig.tagsNotIn
-                  ? {
-                      in: agentConfig.tagsIn || [],
-                      not: agentConfig.tagsNotIn || [],
-                      mode: agentConfig.tagsMode || "custom",
-                    }
-                  : undefined,
-            },
-            dataSource: {
-              dustAPIProjectId: agentConfig.dataSource.dustAPIProjectId,
-              dustAPIDataSourceId: agentConfig.dataSource.dustAPIDataSourceId,
-              connectorProvider: agentConfig.dataSource.connectorProvider,
-              name: agentConfig.dataSource.name,
-            },
-          };
-          return new Ok(resolved);
-        }
-
-        case "dynamic": {
-          // Dynamic configuration
-          // Verify the workspace ID matches the auth
-          if (
-            configInfo.configuration.workspaceId !==
-            auth.getNonNullableWorkspace().sId
-          ) {
-            return new Err(
-              new Error(
-                "Workspace mismatch: configuration workspace " +
-                  `${configInfo.configuration.workspaceId} does not match authenticated workspace.`
-              )
-            );
-          }
-
-          // Fetch the specific data source view by ID
-          const dataSourceView = await DataSourceViewResource.fetchById(
-            auth,
-            configInfo.configuration.dataSourceViewId
-          );
-
-          if (!dataSourceView) {
-            return new Err(
-              new Error(
-                `Data source view not found: ${configInfo.configuration.dataSourceViewId}`
-              )
-            );
-          }
-
-          const dataSource = dataSourceView.dataSource;
-
-          const resolved: ResolvedDataSourceConfiguration = {
-            ...configInfo.configuration,
-            dataSource: {
-              dustAPIProjectId: dataSource.dustAPIProjectId,
-              dustAPIDataSourceId: dataSource.dustAPIDataSourceId,
-              connectorProvider: dataSource.connectorProvider,
-              name: dataSource.name,
-            },
-          };
-          return new Ok(resolved);
-        }
-
-        default:
-          assertNever(configInfo);
-      }
-    },
-    { concurrency: 10 }
-  );
-
-  if (configResults.some((res) => res.isErr())) {
-    return new Err(new Error("Failed to fetch data source configurations."));
-  }
-
-  return new Ok(
-    removeNulls(configResults.map((res) => (res.isOk() ? res.value : null)))
-  );
-}
-
-export function makeDataSourceViewFilter(
-  agentDataSourceConfigurations: ResolvedDataSourceConfiguration[]
-) {
-  return agentDataSourceConfigurations.map(({ dataSource, filter }) => ({
-    data_source_id: dataSource.dustAPIDataSourceId,
-    view_filter: filter.parents?.in ?? [],
-  }));
-}
-
 function getSortDirection(field: "title" | "timestamp"): "asc" | "desc" {
   switch (field) {
     case "title":
@@ -1166,6 +1025,46 @@ function extractDataSourceIdFromNodeId(nodeId: string): string | null {
     return null;
   }
   return nodeId.substring(`${DATA_SOURCE_NODE_ID}-`.length);
+}
+
+export function makeQueryResourceForFind(
+  query?: string,
+  rootNodeId?: string,
+  mimeTypes?: string[],
+  nextPageCursor?: string
+): SearchQueryResourceType {
+  const queryText = query ? ` "${query}"` : " all content";
+  const scope = rootNodeId
+    ? ` under ${rootNodeId}`
+    : " across the entire data sources";
+  const types = mimeTypes?.length
+    ? ` (${mimeTypes.map(renderMimeType).join(", ")} files)`
+    : "";
+  const pagination = nextPageCursor ? " - next page" : "";
+
+  return {
+    mimeType: INTERNAL_MIME_TYPES.TOOL_OUTPUT.DATA_SOURCE_SEARCH_QUERY,
+    text: `Searching for${queryText}${scope}${types}${pagination}.`,
+    uri: "",
+  };
+}
+
+export function makeQueryResourceForList(
+  nodeId: string | null,
+  mimeTypes?: string[],
+  nextPageCursor?: string
+): SearchQueryResourceType {
+  const location = nodeId ? ` within node "${nodeId}"` : " at the root level";
+  const types = mimeTypes?.length
+    ? ` (${mimeTypes.map(renderMimeType).join(", ")} files)`
+    : "";
+  const pagination = nextPageCursor ? " - next page" : "";
+
+  return {
+    mimeType: INTERNAL_MIME_TYPES.TOOL_OUTPUT.DATA_SOURCE_SEARCH_QUERY,
+    text: `Listing content${location}${types}${pagination}.`,
+    uri: "",
+  };
 }
 
 export default createServer;
