+++ front/lib/actions/mcp_internal_actions/servers/utils.ts
@@ -26,7 +26,6 @@ import type {
   CoreAPISearchFilter,
   DataSourceViewType,
   Result,
-  TimeFrame,
 } from "@app/types";
 import { assertNever, Err, Ok } from "@app/types";
 
@@ -391,28 +390,4 @@ export function shouldAutoGenerateTags(
   }
 
   return false;
-}
-
-export function renderRelativeTimeFrameForToolOutput(
-  relativeTimeFrame: TimeFrame | null
-): string {
-  return relativeTimeFrame
-    ? "over the last " +
-        (relativeTimeFrame.duration > 1
-          ? `${relativeTimeFrame.duration} ${relativeTimeFrame.unit}s`
-          : `${relativeTimeFrame.unit}`)
-    : "across all time periods";
-}
-
-export function renderTagsForToolOutput(
-  tagsIn?: string[],
-  tagsNot?: string[]
-): string {
-  const tagsInAsString =
-    tagsIn && tagsIn.length > 0 ? `, with labels ${tagsIn?.join(", ")}` : "";
-  const tagsNotAsString =
-    tagsNot && tagsNot.length > 0
-      ? `, excluding labels ${tagsNot?.join(", ")}`
-      : "";
-  return `${tagsInAsString}${tagsNotAsString}`;
 }
